packages/driver/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stretch3_zmq_driver"]

[project]
name = "stretch3-zmq-driver"
version = "0.0.3"
requires-python = ">=3.12, <3.13"
dependencies = [
    "stretch3-zmq-core",
    "elevenlabs>=2.32.0",
    "pyzmq>=27.1.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "dotenv>=0.9.9",
    "rich>=13.0.0",
    "scipy>=1.17.0",
    "hello-robot-stretch-body==0.7.29",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "ikpy>=3.4.2",
    "graphviz>=0.21",
    "urchin>=0.0.30",
    "websockets==12.0",
    "opencv-python>=4.8,<4.13",
]

[project.scripts]
stretch3-zmq-driver = "stretch3_zmq_driver.__main__:main"

[tool.uv.sources]
stretch3-zmq-core = { workspace = true }
